+++ resources/js/pages/Home.vue
@@ -1,15 +1,95 @@
 <template>
-        <div class="container">
-            <h1>Homepage</h1>
+    <div class="container">
+        <div class="search-bar">
+            <input
+                type="text"
+                v-model="searchText"
+                placeholder="Search for cuisine"
+                @keyup.enter="getRestaurants"
+            />
+            <button @click="getRestaurants">Search</button>
         </div>
+        <h1>Homepage</h1>
+            <Cuisines :cuisines="cuisines"/>
+        
+        <article v-for="restaurant in results" :key="restaurant.id">
+            <h2>{{ restaurant.name }}</h2>
+            <router-link
+                :to="{
+                    name: 'restaurant-detail',
+                    params: { name: restaurant.name }
+                }"
+                >Restaurant Detail</router-link
+            >
+        </article>
+    </div>
 </template>
 
 <script>
+import axios from "axios";
+import Cuisines from "../components/Cuisines.vue";
 export default {
-    name: 'Home',
-}
+    name: "Home",
+    components: {
+        Cuisines,
+    },
+    data() {
+        return {
+            apiURL: "http://127.0.0.1:8000/api/restaurants",
+            restaurants: [],
+            searchText: "",
+            listRestaurant: [],
+            results: []
+        };
+    },
+    created() {
+        this.getRestaurants();
+    },
+    methods: {
+        getRestaurants() {
+            if (this.searchText === "") {
+                axios
+                    .get(this.apiURL)
+                    .then(res => {
+                        this.restaurants = res.data;
+                        this.results = this.restaurants;
+                    })
+                    .catch(err => {
+                        console.log(err);
+                    });
+            } else if (this.searchText !== "") {
+                var filteredRestaurant = this.restaurants.filter(element => {
+                    for (const item of element.cuisines) {
+                        if (item.type.includes(this.searchText))
+                            return item.type;
+
+                    }
+                });
+
+                this.results = filteredRestaurant;
+            }
+        }
+    }
+};
 </script>
 
-<style>
+<style scoped lang="scss">
+.container {
+    height: 80vh;
 
+    .search-bar {
+        display: flex;
+        justify-content: flex-end;
+    }
+
+    h1 {
+        margin: 20px;
+        display: flex;
+        justify-content: center;
+    }
+
+    h2 {
+        margin: 10px 0;
+    }
+}
 </style>
