+++ resources/js/routes.js
@@ -3,7 +3,6 @@ import Vue from "vue";
 import VueRouter from "vue-router";
 
 import Home from "./pages/Home.vue";
-import Restaurants from "./pages/Restaurants.vue";
 import RestaurantDetail from "./pages/RestaurantDetail.vue";
 import NotFound from "./pages/NotFound.vue";
 
@@ -17,11 +16,6 @@ const router = new VueRouter({
             path: "/",
             name: "home",
             component: Home
-        },
-        {
-            path: "/restaurants",
-            name: "restaurants",
-            component: Restaurants
         },
         {
             path: "/restaurant/:name",
